+++ sho.rt/Areas/Admin/Pages/Mapping/Edit.cshtml.cs
@@ -71,7 +71,7 @@ namespace sho.rt.Areas.Admin
             return RedirectToPage("./Index");
         }
 
-        private bool MappingExists(int id)
+        private bool MappingExists(Int64 id)
         {
             return _context.Mapping.Any(e => e.Id == id);
         }
